CHANGE: Discard any waiting time data in which waiting time exceeds the service interval after only one month CODE: Return a 32-bit integer from the get_average_waiting_time method FIX: The wrong waiting times were appended to the waiting time list in stale cases

+++ simhalt.h
@@ -908,7 +908,7 @@ public:
 
 	// Getting and setting average waiting times in minutes
 	// @author: jamespetts
-	uint16 get_average_waiting_time(halthandle_t halt, uint8 category);
+	uint32 get_average_waiting_time(halthandle_t halt, uint8 category);
 
 	void add_waiting_time(uint32 time, halthandle_t halt, uint8 category, bool do_not_reset_month = false);
